requirements.txt
Django
psycopg2
django-environ